Improve control of the communication rate with the NTRIP server
Added the ability to set the rate of communication with the NTRIP Server. Some servers have policies that will lock out the connection if requests are sent too often.

Added a convenience launch file to bring up all 3 nodes together

Changed the defaults for the f9p to a faster but achievable solution rate